Renaming io dir to utils, adjusting io_processing for config based file aliases. Renaming helper_classes to data_classes
